requirements.txt
tox
markdown
factory_boy
djangorestframework
django-debug-toolbar
